+++ trunk/ardumower/drivers.h
@@ -145,8 +145,8 @@ String date2str(date_t date);
 String versionToStr(byte v[]);
 
 // I2C helpers
-void I2CwriteTo(int device, byte address, byte val);
-int I2CreadFrom(int device, byte address, int num, byte buff[], int retryCount = 0);
+void I2CwriteTo(uint8_t device, uint8_t address, uint8_t val);
+int I2CreadFrom(uint8_t device, uint8_t address, uint8_t num, uint8_t buff[], int retryCount = 0);
 
 // rescale to -PI..+PI
 double scalePI(double v);
